Convert to sqlite

+++ priv/repo/migrations/20211112221650_create_subscription.exs
@@ -3,7 +3,7 @@ defmodule YTD.Repo.Migrations.CreateSubscription do
 
   def change do
     create table(:subscription) do
-      add :strava_id, :decimal, unique: true
+      add :strava_id, :bigint, unique: true
       timestamps()
     end
